+++ src/main/java/models/SpringAnalyticSolution.java
@@ -1,11 +1,9 @@
 package models;
 
-import models.AnalyticSolution;
-
 public class SpringAnalyticSolution implements AnalyticSolution {
 
 	@Override
-	public double getPosition(double k, double gamma, double mass, double t) {
-		return Math.exp(-(gamma / (2 * mass)) * t) * Math.cos(Math.sqrt((k / mass) - (gamma * gamma / (4 * mass * mass))) * t);
+	public float getPosition(float k, float gamma, float mass, float t) {
+		return (float) (Math.exp(-(gamma / (2 * mass)) * t) * Math.cos(Math.sqrt((k / mass) - (gamma * gamma / (4 * mass * mass))) * t));
 	}
 }
